requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
boto3==1.18.28
botocore==1.21.28
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-bootstrap-pagination==1.7.1
django-bootstrap4==3.0.1
django-import-export==2.5.0
django-ipware==4.0.0
django-mathfilters==1.0.0
django-modeltranslation==0.17.3
django-storages==1.11.1
et-xmlfile==1.1.0
jmespath==0.10.0
MarkupPy==1.14
mysqlclient==2.0.3
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.2
Pillow==8.3.1
plotly==5.2.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
tenacity==8.0.1
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
